uriparse: use size_t for inet_ntop declaration and fix macro checks
MSDN, mingw-w64, and RFC 2553 specifies the last argument to be size_t
instead of socklen_t that posix specifies.

The macro has been changed to accommodate both MinGW and mingw-w64
for senarios such as if the mingw-w64 version (separate from gcc's)
is lower than the first commit in which the declaration was introduced
in or if the toolchain is targetting a version lower than Vista.

macro check was suggested by @hartwork

Thanks @brechtsanders

Signed-off-by: Christopher Degawa <ccom@randomderp.com>