src/UriCopy.c: Fix copying URIs with empty components
When trying to copy a URI with empty userInfo, portText, query, or fragment
components, the copying process would perform a zero-byte allocation for the
respective component.

Depending on the malloc implementation, this would either return `NULL`,
leading to a bogus `URI_ERROR_MALLOC` being returned from the copying process
or a valid zero-byte allocation that will not be freed by FreeUriMembers,
resulting in a memory leak.