src/UriMemory.c: Remake macros for reallocarray availability
The cause were compile errors with Illumos, where [1] checks
for "defined(_XPG8)" that is set when guard
"(_XOPEN_SOURCE - 0 == 800) || (_POSIX_C_SOURCE - 0 == 202405L)"
evaluates true at [2].

For glibc, support for <2.29 was dropped because 2.29 is >=5 years old,
and 5 years is my default target support window.

For NetBSD I wanted to stop defining _OPENBSD_SOURCE but kept it
because of the 5 years window as of today.

[1] https://github.com/illumos/illumos-gate/blob/ad2626d4276dcc76244ed0fd77efacc93db8d26f/usr/src/head/stdlib.h#L346-L355
[2] https://github.com/illumos/illumos-gate/blob/ad2626d4276dcc76244ed0fd77efacc93db8d26f/usr/src/uts/common/sys/feature_tests.h#L339-L340